requirements.txt
dj-database-url==0.5.0
Django==1.11.6
django-admin-list-filter-dropdown==1.0.2
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-filter==2.1.0
django-smart-selects==1.5.3
et-xmlfile==1.0.1
jdcal==1.4
openpyxl==2.5.14
Pillow==5.2.0
psycopg2==2.7.7
pytz==2018.5
six==1.11.0
transliterate==1.10.2
django-widget-tweaks==1.4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
djangorestframework-datatables==0.4.1
#uwsgi


